Patch 0.1.3: Completed directory refactoring and removed all unused assets
