+++ src/components/aboutme/MainText.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { makeStyles } from '@material-ui/core/styles';
 import {motion} from "framer-motion";
-import Markdown from "../subcomponents/Markdown";
+import Markdown from "../common/subcomponents/Markdown";
 
 const useStyles = makeStyles((theme) => ({
     markdown: {
