+++ src/components/blog/subcomponents/BlogAccordion.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { makeStyles } from '@material-ui/core/styles';
-import Markdown from './Markdown';
+import Markdown from '../../common/subcomponents/Markdown';
 import Card from "@material-ui/core/Card";
 import AccordionSummary from "@material-ui/core/AccordionSummary";
 import AccordionDetails from "@material-ui/core/AccordionDetails";
