+++ src/components/footer/subcomponents/SocialsCard.js
