+++ src/components/profession/ExperienceContent.js
@@ -2,7 +2,7 @@ import React from "react";
 import {makeStyles} from "@material-ui/core/styles";
 import {motion} from "framer-motion";
 import PropTypes from "prop-types";
-import SpringImageCard from "../subcomponents/SpringImageCard";
+import SpringImageCard from "./subcomponents/SpringImageCard";
 
 const useStyles = makeStyles((theme) => ({
     img: {
