+++ src/components/projects/MainRepositoryList.js
@@ -1,8 +1,8 @@
 import React from "react";
 import Grid from "@material-ui/core/Grid";
 import { motion } from "framer-motion";
-import RepositoryCard from "../subcomponents/RepositoryCard";
-import MainRepositoryCard from "../subcomponents/MainRepositoryCard";
+import RepositoryCard from "./subcomponents/RepositoryCard";
+import MainRepositoryCard from "./subcomponents/MainRepositoryCard";
 import Card from "@material-ui/core/Card";
 
 const transition = {
