+++ src/components/projects/RepositoryAccordion.js
@@ -1,6 +1,6 @@
 import React from "react";
 import WithComponentLoading from "../common/WithComponentLoading";
-import RepositoryList from "../subcomponents/RepositoryList";
+import AllRepositoryList from "./AllRepositoryList";
 import Accordion from '@material-ui/core/Accordion';
 import AccordionSummary from '@material-ui/core/AccordionSummary';
 import AccordionDetails from '@material-ui/core/AccordionDetails';
@@ -26,7 +26,7 @@ const accordionVariants = {
 
 function RepositoryAccordion(props) {
     const { listState } = props;
-    const ListLoading = WithComponentLoading(RepositoryList);
+    const ListLoading = WithComponentLoading(AllRepositoryList);
     const classes = useStyles();
 
     return (
