--- src/components/subcomponents/FeaturedPost.js
@@ -1,61 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { makeStyles } from '@material-ui/core/styles';
-import Typography from '@material-ui/core/Typography';
-import Grid from '@material-ui/core/Grid';
-import Card from '@material-ui/core/Card';
-import CardActionArea from '@material-ui/core/CardActionArea';
-import CardContent from '@material-ui/core/CardContent';
-import CardMedia from '@material-ui/core/CardMedia';
-import Hidden from '@material-ui/core/Hidden';
-
-const useStyles = makeStyles({
-    card: {
-        display: 'flex'
-    },
-    cardDetails: {
-        flex: 1,
-    },
-    cardMedia: {
-        width: 160,
-    },
-});
-
-function FeaturedPost(props) {
-    const classes = useStyles();
-    const { post } = props;
-
-    return (
-        <Grid item xs={12} md={6}>
-            <CardActionArea component="a" href="#">
-                <Card className={classes.card}>
-                    <div className={classes.cardDetails}>
-                        <CardContent>
-                            <Typography component="h2" variant="h5" color="primary">
-                                {post.title}
-                            </Typography>
-                            <Typography variant="subtitle1" color="secondary">
-                                {post.date}
-                            </Typography>
-                            <Typography variant="subtitle1" paragraph color="primary">
-                                {post.description}
-                            </Typography>
-                            <Typography variant="subtitle1" color="secondary">
-                                Continue reading...
-                            </Typography>
-                        </CardContent>
-                    </div>
-                    <Hidden xsDown>
-                        <CardMedia className={classes.cardMedia} image={post.image} title={post.imageTitle} />
-                    </Hidden>
-                </Card>
-            </CardActionArea>
-        </Grid>
-    );
-}
-
-export default FeaturedPost;
-
-FeaturedPost.propTypes = {
-    post: PropTypes.object,
-};
