+++ src/pages/ProfessionPage.js
@@ -6,7 +6,7 @@ import workportfolio1 from "../res/professionposts/work-experience1-2.md";
 import subtitle2 from "../res/professionposts/work-experience2.md";
 import workportfolio2 from "../res/professionposts/work-experience2-2.md";
 import educationportfolio1 from "../res/professionposts/education-experience.md";
-import ProfessionGrid from "./blocks/ProfessionGrid";
+import ProfessionGrid from "./grids/ProfessionGrid";
 
 const useStyles = makeStyles((theme) => ({
     root: {
